Move vector types to different headers
