+++ src/include/duckdb/common/vector/flat_vector.hpp
@@ -1,0 +1,74 @@
+//===----------------------------------------------------------------------===//
+//                         DuckDB
+//
+// duckdb/common/vector/map_vector.hpp
+//
+//
+//===----------------------------------------------------------------------===//
+
+#pragma once
+
+#include "duckdb/common/types/vector.hpp"
+
+namespace duckdb {
+
+struct FlatVector {
+	static void VerifyFlatVector(const Vector &vector) {
+#ifdef DUCKDB_DEBUG_NO_SAFETY
+		D_ASSERT(vector.GetVectorType() == VectorType::FLAT_VECTOR);
+#else
+		if (vector.GetVectorType() != VectorType::FLAT_VECTOR) {
+			throw InternalException("Operation requires a flat vector but a non-flat vector was encountered");
+		}
+#endif
+	}
+
+	static inline data_ptr_t GetData(Vector &vector) {
+		return ConstantVector::GetData(vector);
+	}
+	template <class T>
+	static inline const T *GetData(const Vector &vector) {
+		return ConstantVector::GetData<T>(vector);
+	}
+	template <class T>
+	static inline T *GetData(Vector &vector) {
+		return ConstantVector::GetData<T>(vector);
+	}
+	template <class T>
+	static inline const T *GetDataUnsafe(const Vector &vector) {
+		return ConstantVector::GetDataUnsafe<T>(vector);
+	}
+	template <class T>
+	static inline T *GetDataUnsafe(Vector &vector) {
+		return ConstantVector::GetDataUnsafe<T>(vector);
+	}
+	static inline void SetData(Vector &vector, data_ptr_t data) {
+		D_ASSERT(vector.GetVectorType() == VectorType::FLAT_VECTOR);
+		vector.data = data;
+	}
+	template <class T>
+	static inline T GetValue(Vector &vector, idx_t idx) {
+		D_ASSERT(vector.GetVectorType() == VectorType::FLAT_VECTOR);
+		return FlatVector::GetData<T>(vector)[idx];
+	}
+	static inline const ValidityMask &Validity(const Vector &vector) {
+		VerifyFlatVector(vector);
+		return vector.validity;
+	}
+	static inline ValidityMask &Validity(Vector &vector) {
+		VerifyFlatVector(vector);
+		return vector.validity;
+	}
+	static inline void SetValidity(Vector &vector, const ValidityMask &new_validity) {
+		VerifyFlatVector(vector);
+		vector.validity.Initialize(new_validity);
+	}
+	DUCKDB_API static void SetNull(Vector &vector, idx_t idx, bool is_null);
+	static inline bool IsNull(const Vector &vector, idx_t idx) {
+		D_ASSERT(vector.GetVectorType() == VectorType::FLAT_VECTOR);
+		return !vector.validity.RowIsValid(idx);
+	}
+	DUCKDB_API static const SelectionVector *IncrementalSelectionVector();
+};
+
+}
