+++ src/include/duckdb/common/vector/string_vector.hpp
@@ -1,0 +1,43 @@
+//===----------------------------------------------------------------------===//
+//                         DuckDB
+//
+// duckdb/common/vector/map_vector.hpp
+//
+//
+//===----------------------------------------------------------------------===//
+
+#pragma once
+
+#include "duckdb/common/types/vector.hpp"
+
+namespace duckdb {
+
+struct StringVector {
+	//! Add a string to the string heap of the vector (auxiliary data)
+	DUCKDB_API static string_t AddString(Vector &vector, const char *data, idx_t len);
+	//! Add a string or a blob to the string heap of the vector (auxiliary data)
+	//! This function is the same as ::AddString, except the added data does not need to be valid UTF8
+	DUCKDB_API static string_t AddStringOrBlob(Vector &vector, const char *data, idx_t len);
+	//! Add a string to the string heap of the vector (auxiliary data)
+	DUCKDB_API static string_t AddString(Vector &vector, const char *data);
+	//! Add a string to the string heap of the vector (auxiliary data)
+	DUCKDB_API static string_t AddString(Vector &vector, string_t data);
+	//! Add a string to the string heap of the vector (auxiliary data)
+	DUCKDB_API static string_t AddString(Vector &vector, const string &data);
+	//! Add a string or a blob to the string heap of the vector (auxiliary data)
+	//! This function is the same as ::AddString, except the added data does not need to be valid UTF8
+	DUCKDB_API static string_t AddStringOrBlob(Vector &vector, string_t data);
+	//! Allocates an empty string of the specified size, and returns a writable pointer that can be used to store the
+	//! result of an operation
+	DUCKDB_API static string_t EmptyString(Vector &vector, idx_t len);
+	//! Returns a reference to the underlying VectorStringBuffer - throws an error if vector is not of type VARCHAR
+	DUCKDB_API static VectorStringBuffer &GetStringBuffer(Vector &vector);
+	//! Adds a reference to a handle that stores strings of this vector
+	DUCKDB_API static void AddHandle(Vector &vector, BufferHandle handle);
+	//! Adds a reference to an unspecified vector buffer that stores strings of this vector
+	DUCKDB_API static void AddBuffer(Vector &vector, buffer_ptr<VectorBuffer> buffer);
+	//! Add a reference from this vector to the string heap of the provided vector
+	DUCKDB_API static void AddHeapReference(Vector &vector, Vector &other);
+};
+
+}
